Add files via upload
Here I learnt the difference between OLTP vs OLAP and also did a minor project of ETL using RDS on AWS. First we have the data which is not transformed. Then fetched the data from AWS and implement it on google collab. Then again uploaded the transformed data on the server.